# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample server for a water station.",
        "title": "Swagger Example API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "https://poseidon-backend.fly.dev/",
    "basePath": "/",
    "paths": {
...
        "/refill_station_reviews": {
            "get": {
                "description": "Get all refill station reviews",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "refill_station_reviews"
                ],
                "summary": "Show all refill station reviews",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/database.RefillStationReview"
                            }
                        }
                    }
                }
            },
            "put": {
                "description": "Update an existing refill station review",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "refill_station_reviews"
                ],
                "summary": "Update a refill station review",
                "parameters": [
                    {
                        "description": "Refill Station Review",
                        "name": "review",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/database.RefillStationReview"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/database.RefillStationReview"
                        }
                    }
                }
            },
            "post": {
                "description": "Create a new refill station review",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "refill_station_reviews"
                ],
                "summary": "Create a refill station review",
                "parameters": [
                    {
                        "description": "Refill Station Review",
                        "name": "review",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/database.RefillStationReview"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/database.RefillStationReview"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete an existing refill station review",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "refill_station_reviews"
                ],
                "summary": "Delete a refill station review",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Refill Station Review ID",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    }
                }
            }
        },
...
        "/water_transactions": {
            "get": {
                "description": "Get all water transactions",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "water_transactions"
                ],
                "summary": "Show all water transactions",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/database.WaterTransaction"
                            }
                        }
                    }
                }
            },
            "put": {
                "description": "Update an existing water transaction",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "water_transactions"
                ],
                "summary": "Update a water transaction",
                "parameters": [
                    {
                        "description": "Water Transaction",
                        "name": "transaction",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/database.WaterTransaction"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/database.WaterTransaction"
                        }
                    }
                }
            },
            "post": {
                "description": "Create a new water transaction",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "water_transactions"
                ],
                "summary": "Create a water transaction",
                "parameters": [
                    {
                        "description": "Water Transaction",
                        "name": "transaction",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/database.WaterTransaction"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/database.WaterTransaction"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete an existing water transaction",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "water_transactions"
                ],
                "summary": "Delete a water transaction",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Water Transaction ID",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    }
                }
            }
        }
    },
    "definitions": {
...
        "database.RefillStationReview": {
            "type": "object",
            "properties": {
                "accessibility": {
                    "type": "integer"
                },
                "cleanness": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "station_id": {
                    "type": "integer"
                },
                "timestamp": {
                    "type": "string"
                },
                "user_id": {
                    "type": "integer"
                },
                "water_quality": {
                    "type": "integer"
                }
            }
        },
...
        "database.WaterTransaction": {
            "type": "object",
            "properties": {
                "bottle_id": {
                    "type": "integer"
                },
                "guest": {
                    "type": "boolean"
                },
                "id": {
                    "type": "integer"
                },
                "station_id": {
                    "type": "integer"
                },
                "timestamp": {
                    "type": "string"
                },
                "user_id": {
                    "type": "integer"
                },
                "volume": {
                    "type": "integer"
                },
                "water_type": {
                    "type": "string"
                }
            }
        }
    }
}